+++ src/main/java/jse/cptr/NestedDoubleCPointer.java
@@ -36,7 +36,7 @@ public class NestedDoubleCPointer extends AnyCPointer {
      * @return 创建的嵌套指针的 c 指针对象
      */
     @UnsafeJNI("Manual free required")
-    public static NestedDoubleCPointer malloc(int aCount) {
+    public static NestedDoubleCPointer malloc(long aCount) {
         return new NestedDoubleCPointer(malloc_(aCount, TYPE_SIZE));
     }
     /**
@@ -48,7 +48,7 @@ public class NestedDoubleCPointer extends AnyCPointer {
      * @return 创建的嵌套指针的 c 指针对象
      */
     @UnsafeJNI("Manual free required")
-    public static NestedDoubleCPointer calloc(int aCount) {
+    public static NestedDoubleCPointer calloc(long aCount) {
         return new NestedDoubleCPointer(calloc_(aCount, TYPE_SIZE));
     }
     
@@ -109,11 +109,11 @@ public class NestedDoubleCPointer extends AnyCPointer {
      * @param aColNum 需要读取的 aData 的列数
      */
     @UnsafeJNI("Invalid input nrows/ncols may directly result in JVM SIGSEGV")
-    public void fill(double aValue, int aRowNum, int aColNum) {
+    public void fill(double aValue, long aRowNum, long aColNum) {
         if (isNull()) throw new NullPointerException();
         fill1(mPtr, aValue, aRowNum, aColNum);
     }
-    private native static void fill1(long rPtr, double aValue, int aRowNum, int aColNum);
+    private native static void fill1(long rPtr, double aValue, long aRowNum, long aColNum);
     
     /**
      * 将此嵌套 c 指针对应的内存数值写入 jse 的 {@link RowMatrix} 中，认为数据按行排列且每个内部的
@@ -179,7 +179,7 @@ public class NestedDoubleCPointer extends AnyCPointer {
      * @return {@inheritDoc}
      */
     @UnsafeJNI("Invalid input index may directly result in JVM SIGSEGV")
-    @Override public DoubleCPointer getAt(int aIdx) {
+    @Override public DoubleCPointer getAt(long aIdx) {
         if (isNull()) throw new NullPointerException();
         return new DoubleCPointer(getAt_(mPtr, aIdx));
     }
@@ -190,11 +190,11 @@ public class NestedDoubleCPointer extends AnyCPointer {
      * @return 此指针对应的值
      */
     @UnsafeJNI("Invalid input row/col may directly result in JVM SIGSEGV")
-    public double getAt(int aRow, int aCol) {
+    public double getAt(long aRow, long aCol) {
         if (isNull()) throw new NullPointerException();
         return getAt_(mPtr, aRow, aCol);
     }
-    private native static double getAt_(long aPtr, int aRow, int aCol);
+    private native static double getAt_(long aPtr, long aRow, long aCol);
     
     /**
      * 将此嵌套指针当作一个 c 的矩阵，设置内部指定行列的值，即对应 c 中的 {@code ptr[aRow][aCol] = aValue}
@@ -203,18 +203,18 @@ public class NestedDoubleCPointer extends AnyCPointer {
      * @param aValue 需要设置的值
      */
     @UnsafeJNI("Invalid input row/col may directly result in JVM SIGSEGV")
-    public void putAt(int aRow, int aCol, double aValue) {
+    public void putAt(long aRow, long aCol, double aValue) {
         if (isNull()) throw new NullPointerException();
         putAt_(mPtr, aRow, aCol, aValue);
     }
-    native static void putAt_(long aPtr, int aRow, int aCol, double aValue);
+    native static void putAt_(long aPtr, long aRow, long aCol, double aValue);
     
     /**
      * {@inheritDoc}
      * @param aCount {@inheritDoc}
      * @return {@inheritDoc}
      */
-    @Override public NestedDoubleCPointer plus(int aCount) {
+    @Override public NestedDoubleCPointer plus(long aCount) {
         if (isNull()) throw new NullPointerException();
         return new NestedDoubleCPointer(rightShift_(mPtr, aCount));
     }
@@ -223,7 +223,7 @@ public class NestedDoubleCPointer extends AnyCPointer {
      * @param aCount {@inheritDoc}
      * @return {@inheritDoc}
      */
-    @Override public NestedDoubleCPointer minus(int aCount) {
+    @Override public NestedDoubleCPointer minus(long aCount) {
         if (isNull()) throw new NullPointerException();
         return new NestedDoubleCPointer(leftShift_(mPtr, aCount));
     }
